Split to ease modularization
