+++ pyms/flask/metrics/__init__.py
@@ -1,1 +1,1 @@
-from pyms.flask.metrics.metrics import metrics_blueprint, monitor
+from pyms.flask.metrics.flask import metrics_blueprint, monitor
